# geoFields/src/main/resources/static/OpenAPIspec.yaml
openapi: 3.0.3
info:
  title: GeoFields HTTP API
  description: |
    JSON-эндпоинты бэкенда GeoFields (поля, NDVI, менеджер организации).

    **Авторизация:** сессия Spring Security (форма логина). Запросы с браузера отправляйте с `credentials: include` / same-origin cookie.

    **CSRF:** для `POST` используется cookie `XSRF-TOKEN` и заголовок `X-XSRF-TOKEN` (см. Spring `CookieCsrfTokenRepository`).

    Главная `/` отдаёт HTML без данных пользователя; контекст сессии — `GET /api/session/context`.

    HTML-страницы (`/`, `/login`, `/register`, `/org/manager`, `/org/admin`, `/org/agronomist`) как разметка в спецификации не описаны.
  version: 1.4.3
servers:
  - url: http://localhost:8080
    description: Локальный сервер (см. `server.port` в application.properties)
...
paths:
...
  /get_slope_tiles_for_field:
    get:
      tags: [Slope]
      summary: URL тайлов уклонов для одного поля
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: query
          required: true
          schema: { type: string }
          example: "1"
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2024-06-15"
      responses:
        "200":
          description: URL тайлов и фактическая дата
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlopeTilesForFieldResponse"
        "400":
          description: Некорректные параметры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
  /get_slope_trend:
    get:
      tags: [Slope]
      summary: Тренд уклонов по группе полей
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: query
          required: true
          schema:
            type: array
            items: { type: integer }
          style: form
          explode: true
          example: [1, 2]
        - name: start_date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2023-01-01"
        - name: end_date
          in: query
          required: true
          schema: { type: string, format: date }
          example: "2024-01-01"
      responses:
        "200":
          description: Данные для графика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlopeTrendResponse"
        "400":
          description: Некорректные параметры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
  /slope/tiles/{field_id}/{date}/{scene_id}/{z}/{x}/{y}.png:
    get:
      tags: [Slope]
      summary: PNG-тайл уклонов (структурно как NDVI)
      security:
        - sessionCookie: []
      parameters:
        - name: field_id
          in: path
          required: true
          schema: { type: string }
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
        - name: scene_id
          in: path
          required: true
          schema: { type: string }
        - name: z
          in: path
          required: true
          schema: { type: integer }
        - name: x
          in: path
          required: true
          schema: { type: integer }
        - name: y
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: PNG тайл
          content:
            image/png:
              schema:
                type: string
                format: binary
        "204":
          description: Пустой тайл (пока без рендера)
        "401":
          description: Нет сессии / не задана организация
        "404":
          description: Поле не принадлежит организации пользователя
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Сессия после входа через POST /login
  schemas:
...
    SlopeTilesForFieldResponse:
      type: object
      properties:
        url:
          type: string
          description: URL шаблон тайлов уклонов `{z}/{x}/{y}`
          example: "/slope/tiles/12/2024-06-15/SLOPE/{z}/{x}/{y}.png"
        actual_date:
          type: string
          format: date
          example: "2024-06-15"
    SlopeTrendResponse:
      type: object
      properties:
        labels:
          type: array
          items: { type: string }
          example: ["Jan", "Feb", "Mar"]
        data:
          type: array
          items: { type: number, format: double }
          example: [3.2, 3.6, 3.4]
...
    SimpleError:
      type: object
      properties:
        error:
          type: string
          example: "Некорректный field_id"
...